refactored class architecture
